I.11 - Move items to pacakge and polish GildedRose
